refactor: 로그인 UI, /dashboard -> /main route 변경
기존 LoginForm을 모듈식 컴포넌트로 대체하여 유지 관리 편의성을 높이고 관심사를 분리함.

auth flow, hook, middleware
/dashboard -> /main 으로 변경
