+++ apps/web/hooks/useLogin.test.ts
@@ -1,7 +1,7 @@
-import { renderHook, act } from '@testing-library/react';
-import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { useLogin } from './useLogin';
+import { act, renderHook } from '@testing-library/react';
+import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { useAuthStore } from '../stores/auth';
+import { useLogin } from './useLogin';
 
 // next/navigation 모의 처리
 const mockPush = vi.fn();
@@ -14,7 +14,7 @@ const mockLogin = vi.fn();
 const mockSetState = vi.fn();
 
 vi.mock('../stores/auth', () => ({
-    useAuthStore: vi.fn(),
+  useAuthStore: vi.fn(),
 }));
 
 describe('useLogin', () => {
@@ -23,13 +23,13 @@ describe('useLogin', () => {
 
     // useAuthStore 모의 구현
     (useAuthStore as vi.Mock).mockImplementation((selector) => {
-        const state = {
-            login: mockLogin,
-            error: null,
-            isAuthenticated: false,
-            ...mockSetState.mock.calls[0]?.[0],
-        };
-        return selector(state);
+      const state = {
+        login: mockLogin,
+        error: null,
+        isAuthenticated: false,
+        ...mockSetState.mock.calls[0]?.[0],
+      };
+      return selector(state);
     });
 
     // getState 모의 구현
@@ -40,8 +40,12 @@ describe('useLogin', () => {
     const { result } = renderHook(() => useLogin());
 
     act(() => {
-      result.current.onChangeEmail({ target: { value: 'test@email.com' } } as React.ChangeEvent<HTMLInputElement>);
-      result.current.onChangePassword({ target: { value: 'password123' } } as React.ChangeEvent<HTMLInputElement>);
+      result.current.onChangeEmail({
+        target: { value: 'test@email.com' },
+      } as React.ChangeEvent<HTMLInputElement>);
+      result.current.onChangePassword({
+        target: { value: 'password123' },
+      } as React.ChangeEvent<HTMLInputElement>);
     });
 
     expect(result.current.email).toBe('test@email.com');
@@ -52,12 +56,18 @@ describe('useLogin', () => {
     const { result } = renderHook(() => useLogin());
 
     act(() => {
-        result.current.onChangeEmail({ target: { value: 'test@email.com' } } as React.ChangeEvent<HTMLInputElement>);
-        result.current.onChangePassword({ target: { value: 'password123' } } as React.ChangeEvent<HTMLInputElement>);
-      });
+      result.current.onChangeEmail({
+        target: { value: 'test@email.com' },
+      } as React.ChangeEvent<HTMLInputElement>);
+      result.current.onChangePassword({
+        target: { value: 'password123' },
+      } as React.ChangeEvent<HTMLInputElement>);
+    });
 
     await act(async () => {
-      await result.current.onSubmit({ preventDefault: vi.fn() } as unknown as React.FormEvent<HTMLFormElement>);
+      await result.current.onSubmit({
+        preventDefault: vi.fn(),
+      } as unknown as React.FormEvent<HTMLFormElement>);
     });
 
     expect(mockLogin).toHaveBeenCalledWith('test@email.com', 'password123');
@@ -70,10 +80,12 @@ describe('useLogin', () => {
     const { result } = renderHook(() => useLogin());
 
     await act(async () => {
-      await result.current.onSubmit({ preventDefault: vi.fn() } as unknown as React.FormEvent<HTMLFormElement>);
+      await result.current.onSubmit({
+        preventDefault: vi.fn(),
+      } as unknown as React.FormEvent<HTMLFormElement>);
     });
 
-    expect(mockPush).toHaveBeenCalledWith('/dashboard');
+    expect(mockPush).toHaveBeenCalledWith('/main');
   });
 
   it('should not redirect on failed login', async () => {
@@ -83,7 +95,9 @@ describe('useLogin', () => {
     const { result } = renderHook(() => useLogin());
 
     await act(async () => {
-      await result.current.onSubmit({ preventDefault: vi.fn() } as unknown as React.FormEvent<HTMLFormElement>);
+      await result.current.onSubmit({
+        preventDefault: vi.fn(),
+      } as unknown as React.FormEvent<HTMLFormElement>);
     });
 
     expect(mockPush).not.toHaveBeenCalled();
