+++ apps/web/types/auth/index.ts
@@ -1,0 +1,9 @@
+export interface LoginFormProps {
+  email: string;
+  password: string;
+  error?: string | null;
+  onSubmit: (e: React.FormEvent<HTMLFormElement>) => void;
+  onChangeEmail: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onChangePassword: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onSocialLoginClick: () => void;
+}
